backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
httpx>=0.25.0
qdrant-client>=1.6.0
langchain-community>=0.1.0
huggingface-hub>=0.20.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
fastapi-limiter>=0.1.0
pyrate-limiter>=3.0.0
aiofiles>=23.2.1
redis-async>=1.0.0
